gps expression change

+++ src/common/db.h
@@ -22,7 +22,7 @@ int db_destruct();
 int db_isTableCreated(const char* imeiName);
 int db_createGPS(const char* tableName);
 int db_createCGI(const char* tableName);
-int db_saveGPS(const char* imeiName, int timestamp, int lat, int lon, char speed, short course);
+int db_saveGPS(const char* imeiName, int timestamp, double lat, double lon, char speed, short course);
 int db_saveCGI(const char* imeiName, int timestamp, short mcc, short mnc, short lac, char ci[]);
 
 int db_doWithOBJ(void (*func)(const char*, int));
